backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Computer Vision and OCR
opencv-python==4.8.1.78
pillow==10.1.0
pytesseract==0.3.10
pyobjc-framework-Quartz>=12.0  # macOS screen capture

# Audio Processing (for combat vision)
numpy>=1.24.0
scipy>=1.10.0
PyAudio>=0.2.13

# HTTP Client for Riot API
aiohttp==3.9.1

# AI/LLM SDKs
anthropic>=0.40.0
openai==1.3.7

# Data Validation
pydantic==2.5.2

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
